requirements.txt
fastapi
uvicorn
scikit-learn
pickle
pillow
numpy